Refactor: improved code structure and organization
